requirements.txt
dtaidistance>=2.3.2
graphviz>=0.19
joblib>=1.1.0
keras>=2.4.3
matplotlib>=3.3.0
mlxtend>=0.19.0
nltk>=3.6.5
numpy>=1.21.2
pandas>=1.3.1
pydot>=1.4.2
pymystem3>=0.2.0
scikit-learn>=1.0.0
seaborn>=0.11.2
simple_elmo>=0.9.0
tensorflow-gpu
tqdm>=4.61.1
streamlit
stqdm
conda
scipy>=1.2.1
scikeras